Function to inform which species must be added to the mega-tree phylogeny in the insertion process.

Function to inform which species must be added to the mega-tree phylogeny in the insertion process.

whichFishAdd(data)

Arguments

data

A data frame with three column containing the name of species (s), the Family (f) and Order (o). This can be generated with function FishTaxaMaker

Value

A data frame containing a column informing at which level the species in data must be added.

Details

This function can be used in order to known which species that must be added in the insertion process made by FishPhyloMaker.

Examples

# \donttest{
    data("taxon_data_PhyloMaker")
    res_test <- whichFishAdd(data = taxon_data_PhyloMaker)
#> Joining with `by = join_by(Subfamily, GenCode, FamCode)`
#> Joining with `by = join_by(FamCode)`
#> Joining with `by = join_by(Order, Ordnum, Class, ClassNum)`
#> Joining with `by = join_by(Class, ClassNum)`
#> Can't retrieve a `chronogram_mrca` for family Parodontidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Loricariidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Callichthyidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Cichlidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Crenuchidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Gymnotidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Anostomidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Poeciliidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Serrasalmidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Hemiodontidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Distichodontidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Ctenoluciidae because the family is monophyletic.
#> [1] "Ctenoluciidae"
#> Can't retrieve a `chronogram_mrca` for family Bryconidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Iguanodectidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Chilodontidae because the family is monophyletic.
#> [1] "Chilodontidae"
#> Can't retrieve a `chronogram_mrca` for family Gasteropelecidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Chalceidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Citharinidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Cynodontidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Erythrinidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Hepsetidae because the family is monophyletic.
#> [1] "Hepsetidae"
#> Can't retrieve a `chronogram_mrca` for family Prochilodontidae because the family is monophyletic.
#> [1] "Tarumaniidae"
#> Can't retrieve a `chronogram_mrca` for family Aspredinidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Mochokidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Trichomycteridae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Akysidae because the family is monophyletic.
#> [1] "Ailiidae"
#> Can't retrieve a `chronogram_mrca` for family Ictaluridae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Anchariidae because the family is monophyletic.
#> [1] "Anchariidae"
#> Can't retrieve a `chronogram_mrca` for family Plotosidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Astroblepidae because the family is monophyletic.
#> [1] "Astroblepidae"
#> Can't retrieve a `chronogram_mrca` for family Austroglanididae because the family is monophyletic.
#> [1] "Austroglanididae"
#> Can't retrieve a `chronogram_mrca` for family Clariidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Pseudopimelodidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Siluridae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Cetopsidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Chacidae because the family is monophyletic.
#> [1] "Chacidae"
#> Can't retrieve a `chronogram_mrca` for family Cranoglanididae because the family is monophyletic.
#> [1] "Cranoglanididae"
#> Can't retrieve a `chronogram_mrca` for family Diplomystidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Pangasiidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Heteropneustidae because the family is monophyletic.
#> [1] "Heteropneustidae"
#> Can't retrieve a `chronogram_mrca` for family Horabagridae because the family is monophyletic.
#> [1] "Horabagridae"
#> [1] "Kryptoglanidae"
#> Can't retrieve a `chronogram_mrca` for family Lacantuniidae because the family is monophyletic.
#> [1] "Lacantuniidae"
#> Can't retrieve a `chronogram_mrca` for family Malapteruridae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Nematogenyidae because the family is monophyletic.
#> [1] "Nematogenyidae"
#> [1] "Phreatobiidae"
#> Can't retrieve a `chronogram_mrca` for family Scoloplacidae because the family is monophyletic.
#> [1] "Scoloplacidae"
#> Can't retrieve a `chronogram_mrca` for family Pholidichthyidae because the family is monophyletic.
#> [1] "Pholidichthyidae"
#> Can't retrieve a `chronogram_mrca` for family Rhamphichthyidae because the family is monophyletic.
#> [1] "Rhamphichthyidae"
#> [1] "Procatopodidae"
#> Can't retrieve a `chronogram_mrca` for family Goodeidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Anablepidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Rivulidae because the family is monophyletic.
#> [1] "Aphaniidae"
#> Can't retrieve a `chronogram_mrca` for family Nothobranchiidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Aplocheilidae because the family is monophyletic.
#> [1] "Fluviphylacidae"
#> Can't retrieve a `chronogram_mrca` for family Fundulidae because the family is monophyletic.
#> [1] "Pantanodontidae"
#> Can't retrieve a `chronogram_mrca` for family Profundulidae because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for family Valenciidae because the family is monophyletic.
#> [1] "Valenciidae"
#> Can't retrieve a `chronogram_mrca` for order Characiformes because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for order Siluriformes because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for order Cichliformes because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for order Gymnotiformes because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for order Cyprinodontiformes because the family is monophyletic.
#> Can't retrieve a `chronogram_mrca` for order Pholidichthyiformes because the family is monophyletic.
#> [1] "Pholidichthyiformes"
#> Warning: Requested 3027 but only found 3026 species.
#> • Pholidichthyiformes
#> ***************************************************************
#> *                          Note:                              *
#> *    force.ultrametric does not include a formal method to    *
#> *    ultrametricize a tree & should only be used to coerce    *
#> *   a phylogeny that fails is.ultramtric due to rounding --   *
#> *    not as a substitute for formal rate-smoothing methods.   *
#> ***************************************************************
# }