Obtaining fish phylogeny according to a local pool of species

Obtaining fish phylogeny according to a local pool of species

FishPhyloMaker(
  data,
  insert.base.node = FALSE,
  return.insertions = TRUE,
  progress.bar = TRUE
)

Arguments

data
A data frame with three columns containing the name of species (s), the Family (f) and the Order (o). This data frame can be generated with tab_function function.

insert.base.node
Logical argument indicating if the species must be added automatically in the family and order (when needed) nodes. Default is FALSE

return.insertions
Logical, if TRUE (default) the output is a list of length two containing the phylogeny and a dataframe with a column indicating at which level each species was inserted.

progress.bar
Logical argument. If TRUE (default) a progress bar will be shown in console.

Value

A newick object containing the phylogeny with the species in data object. If return.insertions = TRUE the output will be a list of length two containing the newick phylogeny and a data frame equal that provided in data plus a column indicating at which level each species was inserted in the tree.

Examples

# \donttest{
    data("taxon_data_PhyloMaker")
    res_phylo <- FishPhyloMaker(data = taxon_data_PhyloMaker,
    insert.base.node = TRUE, 
    return.insertions = TRUE, 
    progress.bar = TRUE)

#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/species_fb_2104.tsv.bz2 in 1000000 line chunks:

#> Rows: 35331 Columns: 101

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (63): SpecCode, Genus, Species, Author, FBname, PicPreferredName, PicPre...
#> dbl (36): SpeciesRefNo, FamCode, GenCode, SubGenCode, Fresh, Brack, Saltwate...
#> lgl  (2): AmphibiousRef, Complete

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

#> 	...Done! (in 7.712989 secs)

#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/genera_fb_2104.tsv.bz2 in 1000000 line chunks:

#> Rows: 11863 Columns: 64

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr  (35): GenCode, GenName, GenAuthorYear, GenAuthor, GenYear, GenInEx, Gen...
#> dbl  (24): GenRefno, FamCode, DepthRangeShallow, FB_SppNb, CofF_SppNb, FoW_S...
#> lgl   (2): Brackish, Freshwater
#> dttm  (3): Dateentered, Datemodified, Datechecked

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

#> 	...Done! (in 1.679397 secs)

#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/families_fb_2104.tsv.bz2 in 1000000 line chunks:

#> Rows: 1288 Columns: 81

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr  (50): FamCode, SortID, Family, AuthorYear, Qualification, CommonName, F...
#> dbl  (23): Species, SexualDimorphism, FossilRef, ValidGenera, Complete, Aspi...
#> lgl   (5): CAS_LIN, ZooBank_ID, CommonName_Chinese, CommonName_Chinese_u, Co...
#> dttm  (3): DateEntered, DateModified, DateChecked

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

#> 	...Done! (in 0.428932 secs)

#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/orders_fb_2104.tsv.bz2 in 1000000 line chunks:

#> Rows: 113 Columns: 56

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr  (35): Ordnum, Order, Synonym, CommonName, BodyShapeI, Brackish, Freshwa...
#> dbl  (12): OrderRefNo, SortNo, AuthorYear, Marine, FamCount, ClassNum, ComAn...
#> lgl   (7): CAS_LIN, ZooBank_ID, DateChecked, CommonName_Spanish, CommonName_...
#> dttm  (2): DateEntered, DateModified

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

#> 	...Done! (in 0.2209101 secs)

#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/classes_fb_2104.tsv.bz2 in 1000000 line chunks:

#> Rows: 14 Columns: 16

#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (7): Class, CommonName, BodyShapeI, WaterSalinity, SuperClass, Subclass,...
#> dbl (5): ClassNum, ClassRefNo, SortNo, Synonym, SpeciesCount
#> lgl (4): AuthorYear, CAS_LIN, ZooBank_ID, TS

#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

#> 	...Done! (in 0.2640011 secs)

#> Joining, by = c("Subfamily", "GenCode", "FamCode")

#> Joining, by = "FamCode"

#> Joining, by = c("Order", "Ordnum", "Class", "ClassNum")

#> Joining, by = c("Class", "ClassNum")

#> Can't retrieve a `chronogram_mrca` for family Parodontidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Loricariidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Callichthyidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Cichlidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Crenuchidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Gymnotidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Anostomidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Poeciliidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Serrasalmidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Hemiodontidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Distichodontidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Ctenoluciidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Bryconidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Iguanodectidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Chilodontidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Gasteropelecidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Chalceidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Citharinidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Cynodontidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Erythrinidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Hepsetidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Prochilodontidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Aspredinidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Mochokidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Trichomycteridae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Akysidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Ictaluridae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Anchariidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Plotosidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Astroblepidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Austroglanididae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Clariidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Pseudopimelodidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Siluridae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Cetopsidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Chacidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Cranoglanididae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Diplomystidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Pangasiidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Heteropneustidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Horabagridae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Lacantuniidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Malapteruridae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Nematogenyidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Scoloplacidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Pholidichthyidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Rhamphichthyidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Goodeidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Anablepidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Rivulidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Nothobranchiidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Aplocheilidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Fundulidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Profundulidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for family Valenciidae because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for order Characiformes because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for order Siluriformes because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for order Cichliformes because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for order Gymnotiformes because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for order Cyprinodontiformes because the family is monophyletic.

#> Can't retrieve a `chronogram_mrca` for order Pholidichthyiformes because the family is monophyletic.

#> [1] "Pholidichthyiformes"

#> Warning: Requested 3027 but only found 3026 species.
#> * Pholidichthyiformes

#> Warning: Astyanax may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Astyanax may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Astyanax may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Astyanax may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Crenicichla may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Hisonotus may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Hypostomus may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Hypostomus may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Hypostomus may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Hypostomus may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Imparfinis may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Imparfinis may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Loricaria may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Loricaria may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Moenkhausia may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Moenkhausia may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

#> Warning: Rineloricaria may not be monophyletic
#>   attaching to the most inclusive group containing members of this genus

# }